refactor(shared): extract API base types module
